Create common module in client package and move valueControl function
